Fix SA1402 violation in AgendaSectionTemplates.cs
